+++ lighthouserc.js
@@ -8,7 +8,7 @@ module.exports = {
         "assertions": {
           "categories:performance": ["error", {"minScore": 0.43}],
           "categories:accessibility": ["error", {"minScore": 0.98}],
-          "categories:best-practices": ["error", {"minScore": 0.93}],
+          "categories:best-practices": ["error", {"minScore": 0.92}],
           "categories:seo": ["error", {"minScore": 0.78}],
         }
       },