+++ udata_front/theme/gouvfr/datagouv-components/src/components/DescriptionList/DescriptionList.vue
@@ -1,0 +1,8 @@
+<template>
+  <dl class="fr-col-12 fr-col-sm-6 fr-col-md-4 fr-col-lg-3 fr-my-0">
+    <slot></slot>
+  </dl>
+</template>
+
+<script setup lang="ts">
+</script>