+++ udata_front/theme/gouvfr/datagouv-components/src/components/index.ts
@@ -1,0 +1,4 @@
+import OrganizationNameWithCertificate from "./Organization";
+import { ResourceAccordion, Loader as ResourceAccordionLoader } from "./ResourceAccordion";
+
+export { OrganizationNameWithCertificate, ResourceAccordion, ResourceAccordionLoader };