+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix vue runtime errors [#343](https://github.com/etalab/udata-front/pull/343)
 - New featured topic : agriculture [#344](https://github.com/etalab/udata-front/pull/344)
+- Create data.gouv.fr-components NPM package and move some `gouvfr` components to it [#324](https://github.com/etalab/udata-front/pull/324)
 
 ## 3.5.0, 3.4.0 and 3.3.0 (2024-01-23)